pyproject.toml
[project]
name = "cheap-rag"
version = "0.1.0"
description = "CHEAP Metadata RAG System - Phase 1: Core RAG + Embeddings + Vector Search"
authors = [
    {name = "CHEAP Project"}
]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "chromadb>=0.4.18",
    "anthropic>=0.39.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.1",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.3",
    "tqdm>=4.66.1",
    "loguru>=0.7.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "javalang>=0.13.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation>=0.45b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "basedpyright>=1.34.0",
    "nox>=2023.4.22",
    "ipython>=8.17.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cheap_rag"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests/}"
test-cov = "pytest --cov {args:tests/}"
typecheck = "basedpyright {args:src/}"
lint = "ruff check {args:src/}"
format = "ruff format {args:src/}"

[tool.uv.workspace]
members = ["."]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/**/*.py" = ["ARG"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--import-mode=importlib",
]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*"]

[tool.coverage.report]
precision = 2
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
    "\\.\\.\\.",  # Protocol methods
]
